@@ -7,7 +7,10 @@ import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import androidx.viewpager2.widget.ViewPager2
 import com.google.android.material.tabs.TabLayoutMediator
-import com.tangem.*
+import com.tangem.Log
+import com.tangem.LogFormat
+import com.tangem.SessionViewDelegate
+import com.tangem.TangemSdk
 import com.tangem.common.card.FirmwareVersion
 import com.tangem.common.core.Config
 import com.tangem.common.services.secure.SecureStorage
@@ -16,6 +19,7 @@ import com.tangem.tangem_demo.ui.settings.SettingsFragment
 import com.tangem.tangem_demo.ui.viewDelegate.ViewDelegateFragment
 import com.tangem.tangem_sdk_new.DefaultSessionViewDelegate
 import com.tangem.tangem_sdk_new.extensions.createLogger
+import com.tangem.tangem_sdk_new.extensions.initBiometricAuthManager
 import com.tangem.tangem_sdk_new.extensions.initNfcManager
 import com.tangem.tangem_sdk_new.storage.create
 import kotlinx.android.synthetic.main.activity_demo.*
@@ -68,23 +72,26 @@ class DemoActivity : AppCompatActivity() {
             allowUntrustedCards = true
             filter.allowedCardTypes = FirmwareVersion.FirmwareType.values().toList()
         }
+        val secureStorage = SecureStorage.create(this)
         val nfcManager = TangemSdk.initNfcManager(this)
+        val authManager = TangemSdk.initBiometricAuthManager(this, secureStorage)
 
         val viewDelegate = DefaultSessionViewDelegate(nfcManager, nfcManager.reader, this)
         viewDelegate.sdkConfig = config
         this.viewDelegate = viewDelegate
 
-        return DefaultTangemSdk(
-            nfcManager.reader,
-            viewDelegate,
-            SecureStorage.create(this),
-            config,
+        return TangemSdk(
+            reader = nfcManager.reader,
+            viewDelegate = viewDelegate,
+            biometricManager = authManager,
+            secureStorage = secureStorage,
+            config = config,
         )
     }
 
     class ViewPagerAdapter(
         private val fgList: List<Class<out Fragment>>,
-        fgActivity: FragmentActivity
+        fgActivity: FragmentActivity,
     ) : FragmentStateAdapter(fgActivity) {
 
         override fun getItemCount(): Int = fgList.size
