@@ -14,11 +14,10 @@ import javax.crypto.Cipher
 import javax.crypto.spec.IvParameterSpec
 import javax.crypto.spec.SecretKeySpec
 
-
 object CryptoUtils {
 
     fun initCrypto() {
-        Security.insertProviderAt(BouncyCastleProvider(), 1)
+        Security.addProvider(BouncyCastleProvider())
         Security.addProvider(EdDSASecurityProvider())
     }
 
@@ -46,8 +45,10 @@ object CryptoUtils {
      *
      * @return Result of a verification
      */
-    fun verify(publicKey: ByteArray, message: ByteArray, signature: ByteArray,
-               curve: EllipticCurve = EllipticCurve.Secp256k1): Boolean {
+    fun verify(
+        publicKey: ByteArray, message: ByteArray, signature: ByteArray,
+        curve: EllipticCurve = EllipticCurve.Secp256k1,
+    ): Boolean {
         return when (curve) {
             EllipticCurve.Secp256k1 -> Secp256k1.verify(publicKey, message, signature)
             EllipticCurve.Secp256r1 -> Secp256r1.verify(publicKey, message, signature)
@@ -65,7 +66,7 @@ object CryptoUtils {
      */
     fun generatePublicKey(
         privateKeyArray: ByteArray,
-        curve: EllipticCurve = EllipticCurve.Secp256k1
+        curve: EllipticCurve = EllipticCurve.Secp256k1,
     ): ByteArray {
         return when (curve) {
             EllipticCurve.Secp256k1 -> Secp256k1.generatePublicKey(privateKeyArray)
@@ -76,7 +77,7 @@ object CryptoUtils {
 
     fun loadPublicKey(
         publicKey: ByteArray,
-        curve: EllipticCurve = EllipticCurve.Secp256k1
+        curve: EllipticCurve = EllipticCurve.Secp256k1,
     ): PublicKey {
         return when (curve) {
             EllipticCurve.Secp256k1 -> Secp256k1.loadPublicKey(publicKey)
@@ -162,7 +163,3 @@ fun ByteArray.hmacSha512(input: ByteArray): ByteArray {
 
 private const val ENCRYPTION_SPEC_PKCS7 = "AES/CBC/PKCS7PADDING"
 private const val ENCRYPTION_SPEC_NO_PADDING = "AES/CBC/NOPADDING"
-
-
-
-
