@@ -20,7 +20,7 @@ object Secp256k1 {
         val privateKeySpec = ECPrivateKeySpec(BigInteger(1, privateKeyArray), createECSpec())
         val privateKey = createKeyFactory().generatePrivate(privateKeySpec)
 
-        val signatureInstance = Signature.getInstance("SHA256withECDSA")
+        val signatureInstance = Signature.getInstance("SHA256withECDSA", "SC")
         signatureInstance.initSign(privateKey)
         signatureInstance.update(data)
 
@@ -37,7 +37,7 @@ object Secp256k1 {
     }
 
     internal fun verify(publicKey: ByteArray, message: ByteArray, signature: ByteArray): Boolean {
-        val signatureInstance = Signature.getInstance("SHA256withECDSA")
+        val signatureInstance = Signature.getInstance("SHA256withECDSA", "SC")
         val loadedPublicKey = loadPublicKey(publicKey)
         signatureInstance.initVerify(loadedPublicKey)
         signatureInstance.update(message)
