@@ -1,10 +1,14 @@
 package com.tangem.operations.resetcode
 
 import com.tangem.Message
+import com.tangem.TangemSdk
 import com.tangem.common.CompletionResult
 import com.tangem.common.StringsLocator
 import com.tangem.common.UserCodeType
-import com.tangem.common.core.*
+import com.tangem.common.core.CompletionCallback
+import com.tangem.common.core.Config
+import com.tangem.common.core.TangemError
+import com.tangem.common.core.TangemSdkError
 import com.tangem.common.extensions.calculateSha256
 import com.tangem.common.extensions.guard
 import com.tangem.operations.CommandResponse
@@ -12,7 +16,6 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 
 class ResetPinService(
-    private val sessionBuilder: SessionBuilder,
     private val stringsLocator: StringsLocator,
     private val config: Config
 ) {
@@ -100,7 +103,7 @@ class ResetPinService(
         }
 
         val command = GetResetPinTokenCommand()
-        sessionBuilder.build(
+        TangemSdk.makeSession(
             config = config,
             cardId = resetCardId,
             initialMessage = Message(
@@ -126,7 +129,7 @@ class ResetPinService(
             return
         }
         val command = SignResetPinTokenCommand(resetPinCard)
-        sessionBuilder.build(
+        TangemSdk.makeSession(
             config = config,
             initialMessage = Message(
                 header = stringsLocator.getString(
@@ -182,7 +185,7 @@ class ResetPinService(
 
 
         val command = ResetPinTask(confirmationCard, accessCodeUnwrapped, passcodeUnwrapped)
-        sessionBuilder.build(
+        TangemSdk.makeSession(
             config = config,
             cardId = resetPinCard.cardId,
             initialMessage = Message(
